Fix #430 Add Dmitry Ivanov  to contributors

+++ pkg/api/bookmarks.go
@@ -80,6 +80,7 @@ func AddBookmarksFile(inFile, outFile string, bms []pdf.Bookmark, conf *model.Co
 		tmpFile = outFile
 	}
 	if f2, err = os.Create(tmpFile); err != nil {
+		f1.Close()
 		return err
 	}
 
@@ -87,7 +88,9 @@ func AddBookmarksFile(inFile, outFile string, bms []pdf.Bookmark, conf *model.Co
 		if err != nil {
 			f2.Close()
 			f1.Close()
-			os.Remove(tmpFile)
+			if outFile == "" || inFile == outFile {
+				os.Remove(tmpFile)
+			}
 			return
 		}
 		if err = f2.Close(); err != nil {
@@ -97,9 +100,7 @@ func AddBookmarksFile(inFile, outFile string, bms []pdf.Bookmark, conf *model.Co
 			return
 		}
 		if outFile == "" || inFile == outFile {
-			if err = os.Rename(tmpFile, inFile); err != nil {
-				return
-			}
+			err = os.Rename(tmpFile, inFile)
 		}
 	}()
